requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
# amqp==5.0.6
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.3.4
async-timeout==4.0.3
attrs==23.2.0
billiard==3.6.4.0
Brotli==1.1.0
celery==5.2.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click
coverage==5.5
cryptography==41.0.5
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20200713
distro==1.9.0
Django==3.2
django-allauth==0.58.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-countries==7.5.1
django-dbbackup==3.3.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-import-export==2.5.0
django-jazzmin==2.4.5
django-prometheus==2.1.0
django-recaptcha==4.0.0
django-timezone-field==4.1.2
django-utils-six==2.0
django-weasyprint==2.2.2
django-widget-tweaks==1.4.8
django_csp==3.8
djangorestframework==3.12.4
et-xmlfile==1.0.1
exceptiongroup==1.2.0
flake8==3.9.0
# fonttools==4.47.2
frozenlist==1.4.1
future==0.18.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jdcal==1.4.1
kombu
lxml==4.9.3
MarkupPy==1.14
mccabe==0.6.1
mindee==4.1.0
multidict==6.0.5
nordigen==1.3.2
oauthlib==3.2.2
odfpy==1.4.1
openai==0.28.0
openpyxl==3.0.7
packaging==23.2
pikepdf==8.9.0
Pillow==10.1.0
pipdeptree==2.0.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydyf==0.8.0
pyflakes==2.3.1
PyJWT==2.8.0
pyphen==0.14.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.15.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.1
stripe==7.8.1
tablib==3.0.0
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.1.0
vine==5.0.0
wcwidth==0.2.5
weasyprint==60.2
webencodings==0.5.1
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zopfli==0.2.3
